+++ Dynamic_HD_Fortran_Code/src/latlon_hd_and_lake_model/latlon_hd_model_io_mod.f90
@@ -11,16 +11,16 @@ function read_river_parameters(river_params_filename) &
     result(river_parameters)
   character(len = max_name_length) :: river_params_filename
   type(riverparameters), pointer :: river_parameters
-  real, pointer, dimension(:,:) :: rdirs
+  real(dp), pointer, dimension(:,:) :: rdirs
   integer, pointer, dimension(:,:) :: river_reservoir_nums
   integer, pointer, dimension(:,:) :: overland_reservoir_nums
   integer, pointer, dimension(:,:) :: base_reservoir_nums
-  real, pointer, dimension(:,:) :: river_retention_coefficients
-  real, pointer, dimension(:,:) :: overland_retention_coefficients
-  real, pointer, dimension(:,:) :: base_retention_coefficients
+  real(dp), pointer, dimension(:,:) :: river_retention_coefficients
+  real(dp), pointer, dimension(:,:) :: overland_retention_coefficients
+  real(dp), pointer, dimension(:,:) :: base_retention_coefficients
   integer, pointer, dimension(:,:) :: landsea_mask_int
   logical, pointer, dimension(:,:) :: landsea_mask
-  real, allocatable, dimension(:,:) :: temp_real_array
+  real(dp), allocatable, dimension(:,:) :: temp_real_array
   integer, allocatable, dimension(:,:) :: temp_integer_array
   integer, dimension(2) :: dimids
   integer :: ncid,varid
@@ -99,18 +99,18 @@ function load_river_initial_values(hd_start_filename) &
     result(river_prognostic_fields)
   character(len = max_name_length) :: hd_start_filename
   type(riverprognosticfields), pointer :: river_prognostic_fields
-  real, pointer,     dimension(:,:) :: river_inflow
-  real, pointer, dimension(:,:) :: river_inflow_temp
-  real, pointer,     dimension(:,:,:) :: base_flow_reservoirs
-  real, pointer, dimension(:,:) :: base_flow_reservoirs_temp
-  real, pointer,    dimension(:,:,:) :: overland_flow_reservoirs
-  real, pointer, dimension(:,:) :: overland_flow_reservoirs_temp
-  real, pointer,     dimension(:,:,:) :: river_flow_reservoirs
-  real, pointer, dimension(:,:) :: river_flow_reservoirs_temp1
-  real, pointer, dimension(:,:) :: river_flow_reservoirs_temp2
-  real, pointer, dimension(:,:) :: river_flow_reservoirs_temp3
-  real, pointer, dimension(:,:) :: river_flow_reservoirs_temp4
-  real, pointer, dimension(:,:) :: river_flow_reservoirs_temp5
+  real(dp), pointer,     dimension(:,:) :: river_inflow
+  real(dp), pointer, dimension(:,:) :: river_inflow_temp
+  real(dp), pointer,     dimension(:,:,:) :: base_flow_reservoirs
+  real(dp), pointer, dimension(:,:) :: base_flow_reservoirs_temp
+  real(dp), pointer,    dimension(:,:,:) :: overland_flow_reservoirs
+  real(dp), pointer, dimension(:,:) :: overland_flow_reservoirs_temp
+  real(dp), pointer,     dimension(:,:,:) :: river_flow_reservoirs
+  real(dp), pointer, dimension(:,:) :: river_flow_reservoirs_temp1
+  real(dp), pointer, dimension(:,:) :: river_flow_reservoirs_temp2
+  real(dp), pointer, dimension(:,:) :: river_flow_reservoirs_temp3
+  real(dp), pointer, dimension(:,:) :: river_flow_reservoirs_temp4
+  real(dp), pointer, dimension(:,:) :: river_flow_reservoirs_temp5
   integer, dimension(2) :: dimids
   integer :: nlat,nlon
   integer :: ncid,varid
@@ -215,8 +215,8 @@ function load_drainages_fields(drainages_filename,first_timestep,last_timestep,&
     result(drainages)
   character(len = max_name_length) :: drainages_filename
   type(riverparameters), intent(in) :: river_parameters
-  real, pointer, dimension(:,:,:) :: drainages
-  real, pointer, dimension(:,:) :: drainages_on_timeslice
+  real(dp), pointer, dimension(:,:,:) :: drainages
+  real(dp), pointer, dimension(:,:) :: drainages_on_timeslice
   integer :: first_timestep, last_timestep
   integer :: nlat,nlon
   integer :: ncid,varid
@@ -242,8 +242,8 @@ function load_runoff_fields(runoffs_filename,first_timestep,last_timestep, &
                             river_parameters) &
     result(runoffs)
   character(len = max_name_length) :: runoffs_filename
-  real, pointer, dimension(:,:,:) :: runoffs
-  real, pointer, dimension(:,:) :: runoffs_on_timeslice
+  real(dp), pointer, dimension(:,:,:) :: runoffs
+  real(dp), pointer, dimension(:,:) :: runoffs_on_timeslice
   type(riverparameters), intent(in) :: river_parameters
   integer :: nlat,nlon
   integer :: ncid,varid
@@ -268,7 +268,7 @@ end function load_runoff_fields
 
 subroutine write_river_flow_field(river_parameters,river_flow_field,timestep)
   type(riverparameters), pointer,intent(in) :: river_parameters
-  real, pointer, dimension(:,:),intent(in) :: river_flow_field
+  real(dp), pointer, dimension(:,:),intent(in) :: river_flow_field
   integer,intent(in) :: timestep
   integer :: ncid,varid
   integer :: lat_dimid,lon_dimid
