+++ Dynamic_HD_Fortran_Code/src/latlon_hd_and_lake_model/latlon_lake_model_io_mod.f90
@@ -8,7 +8,7 @@ implicit none
 
 character(len = max_name_length) :: lake_params_filename
 character(len = max_name_length) :: lake_start_filename
-real :: lake_retention_coefficient
+real(dp) :: lake_retention_coefficient
 
 contains
 
@@ -53,8 +53,8 @@ function read_lake_parameters(instant_throughflow)&
   type(lakeparameters), pointer :: lake_parameters
   logical, pointer, dimension(:,:) :: lake_centers
   integer, pointer, dimension(:,:) :: lake_centers_int
-  real, pointer, dimension(:,:) :: connection_volume_thresholds
-  real, pointer, dimension(:,:) :: flood_volume_thresholds
+  real(dp), pointer, dimension(:,:) :: connection_volume_thresholds
+  real(dp), pointer, dimension(:,:) :: flood_volume_thresholds
   logical, pointer, dimension(:,:) :: flood_local_redirect
   integer, pointer, dimension(:,:) :: flood_local_redirect_int
   logical, pointer, dimension(:,:) :: connect_local_redirect
@@ -81,7 +81,7 @@ function read_lake_parameters(instant_throughflow)&
   integer, pointer, dimension(:,:) :: additional_connect_redirect_lat_index
   integer, pointer, dimension(:,:) :: additional_connect_redirect_lon_index
   integer, allocatable, dimension(:,:) :: temp_integer_array
-  real,    allocatable, dimension(:,:) :: temp_real_array
+  real(dp),    allocatable, dimension(:,:) :: temp_real_array
   integer :: nlat,nlon
   integer :: nlat_coarse,nlon_coarse
   integer :: ncid
@@ -314,11 +314,11 @@ end function read_lake_parameters
 subroutine load_lake_initial_values(initial_water_to_lake_centers,&
                                     initial_spillover_to_rivers, &
                                     step_length)
-  real, pointer, dimension(:,:), intent(inout) :: initial_water_to_lake_centers
-  real, pointer, dimension(:,:), intent(inout) :: initial_spillover_to_rivers
-  real, pointer, dimension(:,:) :: initial_water_to_lake_centers_temp
-  real, pointer, dimension(:,:) :: initial_spillover_to_rivers_temp
-  real :: step_length
+  real(dp), pointer, dimension(:,:), intent(inout) :: initial_water_to_lake_centers
+  real(dp), pointer, dimension(:,:), intent(inout) :: initial_spillover_to_rivers
+  real(dp), pointer, dimension(:,:) :: initial_water_to_lake_centers_temp
+  real(dp), pointer, dimension(:,:) :: initial_spillover_to_rivers_temp
+  real(dp) :: step_length
   integer :: nlat_coarse,nlon_coarse
   integer :: nlat,nlon
   integer :: ncid,varid,dimid
@@ -354,7 +354,7 @@ end subroutine load_lake_initial_values
 subroutine write_lake_volumes_field(lake_volumes_filename,&
                                   lake_parameters,lake_volumes)
   character(len = max_name_length) :: lake_volumes_filename
-  real, pointer, dimension(:,:), intent(inout) :: lake_volumes
+  real(dp), pointer, dimension(:,:), intent(inout) :: lake_volumes
   type(lakeparameters), pointer :: lake_parameters
   integer :: ncid,varid,lat_dimid,lon_dimid
   integer, dimension(2) :: dimids
@@ -410,7 +410,7 @@ subroutine write_diagnostic_lake_volumes(lake_parameters, &
   character(len = max_name_length) :: filename
   integer :: ncid,varid,lat_dimid,lon_dimid
   integer, dimension(2) :: dimids
-  real, dimension(:,:), pointer :: diagnostic_lake_volume
+  real(dp), dimension(:,:), pointer :: diagnostic_lake_volume
     diagnostic_lake_volume => calculate_diagnostic_lake_volumes(lake_parameters,&
                                                                 lake_prognostics,&
                                                                 lake_fields)
